sample-rag-app/requirements.txt
boto3==1.35.19
botocore==1.35.19
faiss-cpu==1.8.0.post1
jmespath==1.0.1
numpy==1.26.4
packaging==24.1
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
urllib3==2.2.3
